+++ src/Akeneo/Category/back/Domain/ValueObject/Attribute/AttributeCode.php
@@ -14,7 +14,7 @@ class AttributeCode
 {
     public function __construct(private string $code)
     {
-        Assert::notEmpty($code);
+        Assert::stringNotEmpty($code);
     }
 
     public function __toString(): string