+++ src/Akeneo/Connectivity/Connection/front/src/model/Apps/wizard-data.ts
@@ -4,6 +4,7 @@ export interface AppWizardData {
     appName: string;
     appLogo: string;
     appUrl: string | null;
+    appIsCertified: boolean;
     scopeMessages: ScopeMessage[];
     oldScopeMessages: ScopeMessage[] | null;
     authenticationScopes: Array<'email' | 'profile'>;