+++ src/Akeneo/Platform/Bundle/ImportExportBundle/front/src/components/StorageConfigurator/index.ts
@@ -2,4 +2,5 @@ export * from './model';
 export * from './LocalStorageConfigurator';
 export * from './SftpStorageConfigurator';
 export * from './AmazonS3StorageConfigurator';
+export * from './MicrosoftAzureStorageConfigurator';
 export * from './GoogleCloudStorageConfigurator';