+++ src/Akeneo/Platform/Bundle/ImportExportBundle/front/src/components/StorageConfigurator/model.test.ts
@@ -1,10 +1,18 @@
 import {FeatureFlags} from '@akeneo-pim-community/shared';
-import {AmazonS3Storage, GoogleCloudStorage, LocalStorage, SftpStorage} from '../model';
-import {isLocalStorage, isSftpStorage, getStorageConfigurator, isAmazonS3Storage, isGoogleCloudStorage} from './model';
+import {AmazonS3Storage, GoogleCloudStorage, LocalStorage, MicrosoftAzureStorage, SftpStorage} from '../model';
+import {
+  isLocalStorage,
+  isSftpStorage,
+  getStorageConfigurator,
+  isAmazonS3Storage,
+  isMicrosoftAzureStorage,
+  isGoogleCloudStorage,
+} from './model';
 import {LocalStorageConfigurator} from './LocalStorageConfigurator';
 import {SftpStorageConfigurator} from './SftpStorageConfigurator';
 import {AmazonS3StorageConfigurator} from './AmazonS3StorageConfigurator';
 import {GoogleCloudStorageConfigurator} from './GoogleCloudStorageConfigurator';
+import {MicrosoftAzureStorageConfigurator} from "./MicrosoftAzureStorageConfigurator";
 
 const featureFlagCollection = {
   job_automation_local_storage: false,
@@ -40,6 +48,13 @@ const amazonS3Storage: AmazonS3Storage = {
   file_path: '/tmp/test.xlsx',
 };
 
+const microsoftAzureStorage: MicrosoftAzureStorage = {
+  type: 'microsoft_azure',
+  connection_string: 'agagag',
+  container_name: 'ahaha',
+  file_path: '/tmp/test.xlsx',
+};
+
 const googleCloudStorage: GoogleCloudStorage = {
   type: 'google_cloud',
   file_path: '/tmp/test.xlsx',
@@ -67,6 +82,12 @@ test('it says if a storage is an amazon s3 storage', () => {
   expect(isSftpStorage(localStorage)).toBe(false);
 });
 
+test('it says if a storage is a microsoft azure storage', () => {
+  expect(isMicrosoftAzureStorage(microsoftAzureStorage)).toBe(true);
+  expect(isMicrosoftAzureStorage(sftpStorage)).toBe(false);
+  expect(isSftpStorage(localStorage)).toBe(false);
+});
+
 test('it says if a storage is a google cloud storage', () => {
   expect(isGoogleCloudStorage(googleCloudStorage)).toBe(true);
   expect(isGoogleCloudStorage(sftpStorage)).toBe(false);
@@ -84,7 +105,7 @@ test('it returns storage configurator', () => {
 
   expect(getStorageConfigurator('amazon_s3', featureFlags)).toBe(AmazonS3StorageConfigurator);
 
-  // Also add microsoft azure
+  expect(getStorageConfigurator('microsoft_azure', featureFlags)).toBe(MicrosoftAzureStorageConfigurator);
 
   expect(getStorageConfigurator('google_cloud', featureFlags)).toBe(GoogleCloudStorageConfigurator);