+++ src/Akeneo/Platform/Bundle/ImportExportBundle/front/src/components/StorageConfigurator/model.ts
@@ -7,11 +7,13 @@ import {
   Storage,
   StorageType,
   localStorageIsEnabled,
+  MicrosoftAzureStorage,
   GoogleCloudStorage,
 } from '../model';
 import {LocalStorageConfigurator} from './LocalStorageConfigurator';
 import {SftpStorageConfigurator} from './SftpStorageConfigurator';
 import {AmazonS3StorageConfigurator} from './AmazonS3StorageConfigurator';
+import {MicrosoftAzureStorageConfigurator} from './MicrosoftAzureStorageConfigurator';
 import {GoogleCloudStorageConfigurator} from './GoogleCloudStorageConfigurator';
 
 type StorageLoginType = 'password' | 'private_key';
@@ -33,6 +35,7 @@ const STORAGE_CONFIGURATORS: StorageConfiguratorCollection = {
   none: null,
   sftp: SftpStorageConfigurator,
   amazon_s3: AmazonS3StorageConfigurator,
+  microsoft_azure: MicrosoftAzureStorageConfigurator,
   google_cloud: GoogleCloudStorageConfigurator,
 };
 
@@ -78,10 +81,27 @@ const isAmazonS3Storage = (storage: Storage): storage is AmazonS3Storage => {
   return (
     'amazon_s3' === storage.type &&
     'file_path' in storage &&
+    typeof 'file_path' === 'string' &&
     'region' in storage &&
+    typeof 'region' === 'string' &&
     'bucket' in storage &&
+    typeof 'bucket' === 'string' &&
     'key' in storage &&
-    'secret' in storage
+    typeof 'key' === 'string' &&
+    'secret' in storage &&
+    typeof 'secret' === 'string'
+  );
+};
+
+const isMicrosoftAzureStorage = (storage: Storage): storage is MicrosoftAzureStorage => {
+  return (
+    'microsoft_azure' === storage.type &&
+    'file_path' in storage &&
+    typeof 'file_path' === 'string' &&
+    'connection_string' in storage &&
+    typeof 'connection_string' === 'string' &&
+    'container_name' in storage &&
+    typeof 'container_name' === 'string'
   );
 };
 
@@ -104,6 +124,7 @@ export {
   isLocalStorage,
   isSftpStorage,
   isAmazonS3Storage,
+  isMicrosoftAzureStorage,
   isGoogleCloudStorage,
   isValidLoginType,
   getStorageConfigurator,