+++ src/Akeneo/Platform/Bundle/ImportExportBundle/front/src/hooks/useCheckStorageConnection.ts
@@ -1,7 +1,7 @@
 import {useEffect, useState} from 'react';
 import {useRoute} from '@akeneo-pim-community/shared';
-import {AmazonS3Storage, GoogleCloudStorage, SftpStorage} from '../components';
-import {isAmazonS3Storage, isGoogleCloudStorage, isSftpStorage} from '../components/StorageConfigurator';
+import {AmazonS3Storage, MicrosoftAzureStorage, GoogleCloudStorage, SftpStorage} from '../components';
+import {isAmazonS3Storage, isMicrosoftAzureStorage, isGoogleCloudStorage, isSftpStorage} from '../components/StorageConfigurator';
 
 const isSftpConnectionFieldFulfilled = (storage: SftpStorage): boolean => {
   return (
@@ -23,13 +23,17 @@ const isAmazonS3ConnectionFieldFulfilled = (storage: AmazonS3Storage): boolean =
   );
 };
 
+const isMicrosoftAzureConnectionFieldFulfilled = (storage: MicrosoftAzureStorage): boolean => {
+  return '' !== storage.connection_string && '' !== storage.container_name && '' !== storage.file_path;
+};
+
 const isGoogleCloudConnectionFieldFulfilled = (storage: GoogleCloudStorage): boolean => {
   return (
-    '' !== storage.file_path && '' !== storage.project_id && '' !== storage.service_account && '' !== storage.bucket
+      '' !== storage.file_path && '' !== storage.project_id && '' !== storage.service_account && '' !== storage.bucket
   );
 };
 
-const useCheckStorageConnection = (storage: SftpStorage | AmazonS3Storage | GoogleCloudStorage) => {
+const useCheckStorageConnection = (storage: SftpStorage | AmazonS3Storage | MicrosoftAzureStorage | GoogleCloudStorage) => {
   const [isValid, setValid] = useState<boolean | undefined>(undefined);
   const [isChecking, setIsChecking] = useState<boolean>(false);
   const route = useRoute('pimee_job_automation_get_storage_connection_check');
@@ -39,6 +43,7 @@ const useCheckStorageConnection = (storage: SftpStorage | AmazonS3Storage | Goog
     !isValid &&
     ((isSftpStorage(storage) && isSftpConnectionFieldFulfilled(storage)) ||
       (isAmazonS3Storage(storage) && isAmazonS3ConnectionFieldFulfilled(storage)) ||
+      (isMicrosoftAzureStorage(storage) && isMicrosoftAzureConnectionFieldFulfilled(storage)) ||
       (isGoogleCloudStorage(storage) && isGoogleCloudConnectionFieldFulfilled(storage)));
 
   useEffect(() => {